refactor WatchList Component
