+++ src/app/components/WatchList/WatchListHeader.tsx
