+++ src/app/pages/WatchList/WatchList.stories.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import WatchList from './WatchList';
+
+export default {
+  title: 'Page/WatchList',
+  component: WatchList,
+  parameters: {
+    layout: 'fullscreen',
+  },
+};
+
+export const Watch = (): JSX.Element => <WatchList />;
